requirements.txt
streamlit
groq
python-dotenv
beautifulsoup4
requests
